+++ geotrek/maintenance/forms.py
@@ -90,6 +90,7 @@ class InterventionForm(CommonForm):
             'height',
             'stake',
             'project',
+            'contractors',
             'access',
             'description',
             'material_cost',
@@ -105,7 +106,7 @@ class InterventionForm(CommonForm):
         model = Intervention
         fields = CommonForm.Meta.fields + \
             ['structure', 'name', 'begin_date', 'end_date', 'status', 'disorders', 'type', 'description', 'subcontracting', 'length', 'width',
-             'height', 'stake', 'project', 'access', 'material_cost', 'heliport_cost', 'contractor_cost', 'workforce_cost', 'topology']
+             'height', 'stake', 'project', 'contractors', 'access', 'material_cost', 'heliport_cost', 'contractor_cost', 'workforce_cost', 'topology']
 
     def __init__(self, *args, target_type=None, target_id=None, **kwargs):
         super().__init__(*args, **kwargs)
@@ -205,9 +206,7 @@ class ProjectForm(CommonForm):
                     'constraint',
                     'global_cost',
                     'comments',
-
-                    css_class="span6"),
-                Div('project_owner',
+                    'project_owner',
                     'project_manager',
                     'contractors',
                     Fieldset(_("Fundings")),